% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\docType{data}
\name{mycolors}
\alias{mycolors}
\title{Mycolors Color Palette}
\format{
An object of class \code{character} of length 14.
}
\usage{
mycolors
}
\description{
A palette of colors designed for visually appealing plots. The \code{mycolors} palette
includes a range of colors suitable for a variety of plotting needs. With lineages names.
}
\keyword{datasets}
