% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\docType{data}
\name{pathogenomics}
\alias{pathogenomics}
\title{PathoGenOmics Color Palette}
\format{
A character vector of 8 hex color codes.
}
\source{
Color palette designed by the PathoGenOmics Lab.
}
\usage{
pathogenomics
}
\description{
A palette of 8 colors from the PathoGenOmics Lab theme.
}
\keyword{datasets}
