% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\docType{data}
\name{PGO}
\alias{PGO}
\alias{pathogenomics}
\title{ClassicTB Color Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
pathogenomics
}
\description{
This palette provides a set of colors derived from the \code{PGO} theme,
suitable for creating distinctive and attractive visualizations.
}
\keyword{datasets}
