% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{plot_tb_cladogram}
\alias{plot_tb_cladogram}
\title{Plot a Phylogenetic Cladogram with TB Lineage Colors}
\usage{
plot_tb_cladogram(newick_text)
}
\arguments{
\item{newick_text}{A character string with the tree in Newick format.}
}
\value{
A ggplot object representing the phylogenetic cladogram.
}
\description{
Visualizes a phylogenetic tree as a cladogram, coloring tips with the \code{mycolors} palette.
}
\examples{
\dontrun{
tree_text <- "(L8,((L1,(L7,(L4,(L2,L3)))),(L5,((A2,(A3,A4)),(A1,(L10,(L6,L9)))))));"
plot_tb_cladogram(tree_text)
}
}
