% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\docType{data}
\name{classicTB}
\alias{classicTB}
\title{ClassicTB Color Palette}
\format{
An object of class \code{character} of length 14.
}
\usage{
classicTB
}
\description{
This palette provides a set of colors derived from the \code{classicTB} theme,
suitable for creating distinctive and attractive visualizations.Without lineages names.
}
\keyword{datasets}
