% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\docType{data}
\name{mycolors}
\alias{mycolors}
\title{Mycolors Color Palette}
\format{
A character vector of 14 hex color codes, named with lineage identifiers.
}
\source{
Color palette designed by the PathoGenOmics Lab.
}
\usage{
mycolors
}
\description{
A named vector of 14 colors designed for visualizing Mycobacterium tuberculosis lineages.
}
\keyword{datasets}
