% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_color_classicTB}
\alias{scale_color_classicTB}
\title{Scale Color for ggplot2 Using mycolors Palette}
\usage{
scale_color_classicTB()
}
\description{
This function provides a color scale for ggplot2 plots, using the custom classicTB palette.
}
