% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_color_classicTB}
\alias{scale_color_classicTB}
\title{Scale Color for ggplot2 Using classicTB Palette}
\usage{
scale_color_classicTB()
}
\value{
A ggplot2 scale object.
}
\description{
Applies the \code{classicTB} palette to the color aesthetic in a ggplot.
}
