% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_color_mycolors}
\alias{scale_color_mycolors}
\title{Scale Color for ggplot2 Using mycolors Palette}
\usage{
scale_color_mycolors()
}
\description{
This function provides a color scale for ggplot2 plots, using the custom mycolors palette with lineage names.
}
