% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_color_mycolors}
\alias{scale_color_mycolors}
\title{Scale Color for ggplot2 Using mycolors Palette}
\usage{
scale_color_mycolors()
}
\value{
A ggplot2 scale object.
}
\description{
Applies the \code{mycolors} palette to the color aesthetic in a ggplot.
}
