% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_fill_classicTB}
\alias{scale_fill_classicTB}
\title{Scale Fill for ggplot2 Using mycolors Palette}
\usage{
scale_fill_classicTB()
}
\description{
This function provides a fill scale for ggplot2 plots, using the custom classicTB palette.
}
