% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_fill_classicTB}
\alias{scale_fill_classicTB}
\title{Scale Fill for ggplot2 Using classicTB Palette}
\usage{
scale_fill_classicTB()
}
\value{
A ggplot2 scale object.
}
\description{
Applies the \code{classicTB} palette to the fill aesthetic in a ggplot.
}
