% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{scale_fill_mycolors}
\alias{scale_fill_mycolors}
\title{Scale Fill for ggplot2 Using mycolors Palette}
\usage{
scale_fill_mycolors()
}
\value{
A ggplot2 scale object.
}
\description{
Applies the \code{mycolors} palette to the fill aesthetic in a ggplot.
}
