% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTB.R
\name{view_palette}
\alias{view_palette}
\title{Display a color palette}
\usage{
view_palette(palette_name = "mycolors")
}
\arguments{
\item{palette_name}{The name of the palette to display ("mycolors", "classicTB", or "pathogenomics").}
}
\value{
A ggplot object showing the colors of the chosen palette.
}
\description{
Generates a ggplot visualization of a specified package palette.
}
\examples{
view_palette("mycolors")
view_palette("classicTB")
}
